Move HeterogeneousDevice(Id) definitions to their own header
